+++ dolby_vision/src/rpu/extension_metadata/blocks/level1.rs
@@ -4,6 +4,8 @@ use bitvec_helpers::{bitvec_reader::BitVecReader, bitvec_writer::BitVecWriter};
 #[cfg(feature = "serde_feature")]
 use serde::{Deserialize, Serialize};
 
+use crate::rpu::vdr_dm_data::CmVersion;
+
 use super::{ExtMetadataBlock, ExtMetadataBlockInfo};
 
 /// cbindgen:ignore
@@ -14,6 +16,7 @@ pub const L1_MAX_PQ_MIN_VALUE: u16 = 2081;
 pub const L1_MAX_PQ_MAX_VALUE: u16 = 4095;
 /// cbindgen:ignore
 pub const L1_AVG_PQ_MIN_VALUE: u16 = 819;
+pub const L1_AVG_PQ_MIN_VALUE_CMV40: u16 = 1229;
 
 /// Statistical analysis of the frame: min, max, avg brightness.
 #[repr(C)]
@@ -52,11 +55,7 @@ impl ExtMetadataBlockLevel1 {
         Ok(())
     }
 
-    pub fn from_stats(min_pq: u16, max_pq: u16, avg_pq: u16) -> ExtMetadataBlockLevel1 {
-        let min_pq = min_pq.clamp(0, L1_MIN_PQ_MAX_VALUE);
-        let max_pq = max_pq.clamp(L1_MAX_PQ_MIN_VALUE, L1_MAX_PQ_MAX_VALUE);
-        let avg_pq = avg_pq.clamp(L1_AVG_PQ_MIN_VALUE, max_pq - 1);
-
+    pub fn new(min_pq: u16, max_pq: u16, avg_pq: u16) -> ExtMetadataBlockLevel1 {
         ExtMetadataBlockLevel1 {
             min_pq,
             max_pq,
@@ -64,10 +63,48 @@ impl ExtMetadataBlockLevel1 {
         }
     }
 
-    pub fn clamp_values(&mut self) {
+    fn clamp_values_int(&mut self, cm_version: CmVersion) {
+        let avg_min_value = match cm_version {
+            CmVersion::V29 => L1_AVG_PQ_MIN_VALUE,
+            CmVersion::V40 => L1_AVG_PQ_MIN_VALUE_CMV40,
+        };
+
         self.min_pq = self.min_pq.clamp(0, L1_MIN_PQ_MAX_VALUE);
         self.max_pq = self.max_pq.clamp(L1_MAX_PQ_MIN_VALUE, L1_MAX_PQ_MAX_VALUE);
-        self.avg_pq = self.avg_pq.clamp(L1_AVG_PQ_MIN_VALUE, self.max_pq - 1);
+        self.avg_pq = self.avg_pq.clamp(avg_min_value, self.max_pq - 1);
+    }
+
+    // Returns a L1 metadata block clamped to valid values
+    pub fn from_stats_cm_version(
+        min_pq: u16,
+        max_pq: u16,
+        avg_pq: u16,
+        cm_version: CmVersion,
+    ) -> ExtMetadataBlockLevel1 {
+        let mut block = Self::new(min_pq, max_pq, avg_pq);
+        block.clamp_values_int(cm_version);
+
+        block
+    }
+
+    pub fn clamp_values_cm_version(&mut self, cm_version: CmVersion) {
+        self.clamp_values_int(cm_version);
+    }
+
+    #[deprecated(
+        since = "1.7.1",
+        note = "Replaced by `from_stats_cm_version` with CmVersion::V29"
+    )]
+    pub fn from_stats(min_pq: u16, max_pq: u16, avg_pq: u16) -> ExtMetadataBlockLevel1 {
+        Self::from_stats_cm_version(min_pq, max_pq, avg_pq, CmVersion::V29)
+    }
+
+    #[deprecated(
+        since = "1.7.1",
+        note = "Replaced by `clamp_values_cm_version` with CmVersion::V29"
+    )]
+    pub fn clamp_values(&mut self) {
+        self.clamp_values_int(CmVersion::V29);
     }
 }